finish three page

+++ MyApplication/app/src/main/java/com/example/myapplication/Activities/MainActivity.java
@@ -38,11 +38,11 @@ public class MainActivity extends AppCompatActivity {
         sliderItems.add(new SliderItems(R.drawable.wide1));
         sliderItems.add(new SliderItems(R.drawable.wide3));
 
-    viewPager2.setAdapter(new SliderAdapters(sliderItems, viewPager2));
-    viewPager2.setClipToPadding(false);
-    viewPager2.setClipChildren(false);
-    viewPager2.setOffscreenPageLimit(3);
-    viewPager2.getChildAt(0).setOverScrollMode(RecyclerView.OVER_SCROLL_NEVER);
+        viewPager2.setAdapter(new SliderAdapters(sliderItems, viewPager2));
+        viewPager2.setClipToPadding(false);
+        viewPager2.setClipChildren(false);
+        viewPager2.setOffscreenPageLimit(3);
+        viewPager2.getChildAt(0).setOverScrollMode(RecyclerView.OVER_SCROLL_NEVER);
 
         CompositePageTransformer compositePageTransformer=new CompositePageTransformer();
         compositePageTransformer.addTransformer(new MarginPageTransformer(40));
@@ -50,7 +50,7 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void transformPage(@NonNull View page, float position) {
                 float r=1-Math.abs(position);
-                page.setScaleY(0.85+r*0.15f);
+                page.setScaleY((float) (0.85+r*0.15f));
             }
         });
